requirements.txt
mysql-connector == 2.1.6
pandas == 0.23.4
zbar == 0.10
Pillow == 5.3.0